Moved simd_view_impl class into a separate header file
